Allow lint not compatible with this version.
This lint only exists in the nightly version of rustdoc and the
version of this library that we have is quite old, so it's not clean
for this lint. Neither are most projects, but this is the only one
that forbids a lint class that includes it.

Allow the lint for now. Updating the project to ToT will also fix the
problem (upstream disabled *all* forbid/deny specifically to work with
nightly
https://github.com/briansmith/untrusted/commit/f934f697ff75c10ab3691ed591d1d93a3de3580f).

Test: m rustdoc
Bug: None
Change-Id: If6a30981bd747906cfe10064df7c02329b269908